requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.3.5
pydantic==2.4.2
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.1
